virtio: Add a shutdown() driver method
Some virtio drivers may need to reset their device's state on reboot.
The virtio-iommu device needs to restore the bypass state before handing
control back to the firmware. Otherwise if bypass is left disabled
firmware might not be able to get the bootloader and kernel from storage
when rebooting the system.

The shutdown() method is left separate from the remove() method (which
is not called on system reboot) because some virtio drivers may not
support being called on reboot. For example the 9p driver's remove()
currently hangs if it is called on reboot.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Change-Id: Idcf059128edb7c067dd8887fbbb619d9082a070e