rm usage of "-k vmlinux" in perf-xuantie-ntrace.txt
This option is supported by perf default, but the running OS may
modify any .text code. So, if we want to analyze pc ranges and instructions
for a risc-v ntrace message, we should dump the running OS as the
trace message is recording the running OS, not the vmlinux generated
by buildroot.

Change-Id: Iee8a4c765d2df5406f4d59a536f7e76d494025e3
Signed-off-by: Jiangshuai Li <jiangshuai_li@linux.alibaba.com>