selftest: livepatch: Encapsulate cleanup commands within a subshell 
In the shell, || is logical operators that execute commands based
on the exit status of the previous command. In the previous code
logic, only disable_lp $MOD_LIVEPATCH was conditionally executed
based on its return value. However, regardless of the return value
of loop_until, the subsequent unload and die commands were
unconditionally executed. This modification places the cleanup
logic within a subshell, which is conditionally executed based
on the return value of loop_until.

Change-Id: Ide0d365d8b62a20e34c7e9d66ec77f3c3a2d1c6a
Signed-off-by: Mao Han <maohan.mh@alibaba-inc.com>