RISC-V: KVM: Fixup guest tpe usage broken
According the spec:

When V=1, both vsstatus.FS and the HS-level sstatus.FS are in
effect. Attempts to execute a floating-point instruction when
either field is 0 (Off) raise an illegal-instruction exception.
Modifying the floating-point state when V=1 causes both fields
to be set to 3 (Dirty).

So, we can't force SR_MS zero in hsstatus csr.

Signed-off-by: Guo Ren (Alibaba DAMO Academy) <guoren.gr@alibaba-inc.com>
Change-Id: I6da4fb5b56d6467c3892b935681ea61f2ce7534c